# feat(emptyabletoptr): Optimize the performance of EmptyableToPtr()